Applied patch by  Uffe Jakobsen: ioutil.c functions does not handle filesystem links right
This problem was found using SDL build on linux:
ioutil.c functions treats any symbolic-link as a file - no matter if the sym-link is actually for a directory.

How to test (under SDL): make a symbolic link that points to a directory and try to navigate to the directory using the sym-link. What happends is that vice will try to open the sym-links as if it was a file.

This is only a problem for platforms that define feature macro _DIRENT_HAVE_D_TYPE