Gtk3: Fix statusbar CRT controls checkbox
The checkbox was added at the same place in a GtkGrid as a separator, leading
to a blank (but toggle-event ware) space. The "toggle CRT controls" item in
the menu has been removed.
Next up is also adding a Mixer controls checkbox to the statusbar, below the
CRT checkbox, and adding a vertical separator after those two controls to make
the statusbar look nicer and more consistent.