Remove VICEDIR, LIBDIR, DOCDIR etc, only keep VICE_DATADIR and VICE_DOCDIR
No need for those old vars nor the new VICE_[ROOT|BIN]DIR I introduced. We have two situations: the Unix way where $prefix determines the root install dir, which can be tweaked via configure, or we have the Windows/MacOS 'stuff-all-in-a-single-dir' approach, in which the dirs are fixed relative paths.