Fix bug #1237 by disallowing to overwrite existing files.
The original problem was with unpatched PET 2001 ROMs which LOAD with SA=1,
which means to the drive that it is opening a file for WRITE. The fsdevice
happily started to overwrite existing files, which is bad.

Disadvantage: now support for SAVE-@ becomes more needed, and it seems that 
it isn't currently supported in the fsdevice.