Refining the approach for bug #1237, because sometimes overwriting in the fsdevice is convenient, and the error case is rare (old PET ROMs and disabled ROM patch).
Created a command line option +/-fsoverwrite and resource FSDeviceOverwrite
(default: +fsoverwrite: no overwrite).

Because it is strange if rawfile.c would look at an option owned by the
fsdevice, I added FILEIO_COMMAND_OVERWRITE which will overwrite existing
files (but the existing FILEIO_COMMAND_WRITE will not). Then the fsdevice
will use the OVERWRITE variant when OPEN"@:filename" is used, or if the
FSDeviceOverwrite is on.

c1541 will use OVERWRITE in all cases where it writes in the file system.
The monitor goes around emulation code in all cases when it accesses the file
system, so it is unaffected.