Gtk3: Default cart: change labels back into GtkEntry
It now look like the entries will be editable at some point, but it's prefered over using labels that don't properly indicate a resource can be set. Also when we start supporting theming, getting any decent colors out of the theme to render the label will be very hard and confusing.