Add window icons to SDL 1.2 builds
Only tested on Debian for now, needs testing on Windows. I'm not sure about the placement of the call to sdl_ui_set_window_icon(), SDL_WM_SetIcon() needs to be called before the first call to SDL_SetVideoMode(), which is what happens now, but perhaps the call can be deferred. And it probably doesn't make sense to set a window icon for fullscreen mode.