Gtk3: Use monospace font and padding of track number for drive widgets, should fix bug #1585
I had to remove the dash between the <device>:<drive> and <side>:<track>.<halftrack> strings to reduce the extra space that was triggered by using a monospace font. I've set the font size to 90%, 75% takes less space but looks off against the rest of the statusbar.
With 4 (dual-drive) devices enabled there's hardly any space left for the message/recording status widget, so we'll probably have to move the message widget onto its own line, perhaps with a slightly smaller (80%-90%) font to reduce the extra vertical screen real estate taken for that.