Gtk3: Statusbar: Make layout code more flexible
The status bar special widgets such as fps/cpu, drive, tape, joystick and volume can now more easily be arranged using statusbar_append() and statusbar_append_end(). The top half of the status bar is now a single row of a varying amount of widgets, where sub-widgets such a drive or tape widgets are first put in a grid before being pushed onto the top row.
The code accessing these widgets no longer needs nor should concern itself with the location of these widgets on the status bar, widgets are either referenced by a pointer in a status bar struct or via a function.

The layout is now basically:
* One row with special widgets
* One row with messages and recording
* One row with keyboard debugging information (hidden by default)

Note: on Windows and MacOS the volume widget is currently still disabled due to the emulated display rendering over the volume slider popup.