Gtk3: Statusbar: use proper accelerator labels for the joysticks popup menu
Don't use hard-coded accelerator strings, inspect the current menu items for the proper accelerators. This has the benefit of also showing the proper accelerators when the user changes the hotkeys.
The code setting the 'live' accelerators should probably be moved into a more common file and renamed so it can be used for other popup menus.