Archdep: de-archdep-ify printerdrv/output-text.c
Make fork_coproc() handle the archdep side of forking/stream redirection, on systems where this isn't implemented it'll log an error and return -1. Should work on Unix, MacOS and Windows; maybe on Haiku but that requires testing. Introduce archdep_close() and archdep_fdopen() to wrap the arch-specific close(2) and fdopen(2) functions.