Gtk3: CRT controls: disable spin buttons
The CRT controls in the settings dialog (Display -> CRT) still had spin buttons in addition to the sliders, for fine-grained control. They made the UI convoluted and confusing so they're now gone. If a user wants better control over the CRT settings they can use `resset` in the monitor, edit vicerc or use the command line.