Fix bug #1725.
In run_pending_alarms(), we call alarm callbacks that are supposed to
have run already but have not, because it was during the current
instruction. So we call them to catch up.
However, in viacore_store() there is an offset applied to a local
version of the clock: `rclk = *(via_context->clk_ptr) -
via_context->write_offset;` with the comment /* stores have a one-cycle
offset if CLK++ happens before store */.
Now with the catch-up execution of alarms, it was using this changed
clock to call the alarm callbacks. In particular it used this clock
value to generate the offset argument to the alarm callbacks.
The CRTC alarm callback crtc_raster_draw_alarm_handler() used this
offset and the global cpu clock to calculate the effective time, but
because its offset value wasn't properly related to that clock, the
result was bogus some of the time. So from time to time (when writes
were happing to the VIA) it could think that the frame suddenly lasted a
different number of cpu clocks.
Teaching the catchup function about the time offset helped.