Fix speeddos+ cable.
This was broken in r42123 when rearranging the code for CB1 edge
detection. For some reason, limiting this to actual changes in the input
breaks this use case (which seems a bug in itself).
You can see if the parallel cable is really used: when the LOAD command
prints start and end addresses.