Fix some C99-compliance issues
The goal is to be able to compile with -std=c99, but the code contains quite
a few uses of non-standard functions. Using -std=c99 also has the effect of
disabling POSIX/BSD functions and types in the C library (at least with glibc),
so we'll have to use macros like _BSD_SOURCE to enable certain functions when
using -std=c99.
For now -std=c99 is commented out in configure.ac until we find a clean way to
enforce C99 and have POSIX C library extensions.