Gtk3: UI actions: only push actions onto the UI thread that require it
Add member `uithread` (bool) to the ui_action_map_t object that indicates if
an action should be pushed onto the UI thread. If either `uithread` or `dialog`
is true the action handler will be executed on the UI thread, otherwise the
handler will be executed on the calling thread.