Gtk3: Make handler_accelerator() return `TRUE` to avoid hotkeys ending up in the emulated machine
Although the prototype for a closure callback specifies it returning nothing,
we need to return `TRUE` because its being used in the context of an accelarator,
which needs to return gboolean. Returning `TRUE` signals the key has been
consumed and thus the key won't end up in the emulated machine.