Superpet Super-OS/9 : small change in resetting to banked memory mapping.
Reading
https://forum.vcfed.org/index.php?threads/commodore-8032-add-on-board-superpet.49423/post-599663
I realised that it was possible that the "FIRQ disable" bit might indeed just
inhibit the FIRQ, but still cause the memory map to be reset from flat to
banked.

In practice this makes little difference, since
- I myself have never seen the diagnostic that this case ever occurs
- when there is no FIRQ the SYNC instruction never completes, so it doesn't
  really matter what the memory mapping is.

Yet, looking at the hand-drawn schematic at
http://mikenaberezny.com/wp-content/uploads/2009/11/superos9-mmu-schematic-r2.jpg
it may well be that this change represent reality better, so I'm changing it.