Github Actions: Add building Debian package on commit (WIP)
Add an action to build a binary deb package of the Gtk3 UI for amd64.
Currently the package should install into /usr. The vice.pdf file is currently
not included.
The resulting package has not been tested at all, so do not use on a system
you like to keep intact =)